put on mock server

+++ config/setup.js
@@ -1,5 +1,13 @@
 import 'dotenv/config'
 import {logIn, config} from '../helpers';
+import {start} from './server'
+
+const baseUrl = process.env.BASE_URL
+const port = process.env.PORT
+const isMock = baseUrl.includes(`localhost:${port}`)
+
+if(isMock)
+  start(port)
 
 before(async  () =>{
   const response = await logIn(process.env.LOGIN,process.env.PASSWORD)
@@ -7,5 +15,6 @@ before(async  () =>{
 })
 
 after(async  () =>{
+  if (!isMock)
   await config.delete()
 })
